Izmenjen typo u upitu

+++ backend/src/main/java/etf/bg/diplomski/teacher/TeacherToSubjectCountDTO.java
@@ -1,4 +1,4 @@
 package etf.bg.diplomski.teacher;
 
-public record TeacherToSubjectCountDTO(String subjectName, int teacherCount) {
+public record TeacherToSubjectCountDTO(String subjectName, long teacherCount) {
 }
